+++ packages/openfort-react/src/hooks/openfort/useConnectWithSiwe.ts
@@ -6,13 +6,19 @@ import { createSIWEMessage } from "../../siwe/create-siwe-message";
 
 import { signMessage } from '@wagmi/core';
 import { useCallback } from "react";
+import { isMobile } from "../../utils";
+import { useWallet } from "../../wallets/useWallets";
 
+// This hook assumes wagmi is already connected to a wallet
+// It will use the connected wallet to sign the SIWE message and authenticate with Openfort
+// If there is a user already, it will link the wallet to the user
 export function useConnectWithSiwe() {
   const { client, user, updateUser } = useOpenfortCore();
   const { log } = useOpenfort();
   const { address, connector } = useAccount();
   const chainId = useChainId();
   const config = useConfig();
+  const wallet = useWallet(connector?.id || "");
 
   const connectWithSiwe = useCallback(async ({
     onError,
@@ -20,7 +26,7 @@ export function useConnectWithSiwe() {
   }: {
     onError?: (error: string, status?: number) => void,
     onConnect?: () => void,
-  }) => {
+  } = {}) => {
     const connectorType = connector?.type;
     const walletClientType = connector?.id;
 
@@ -33,6 +39,7 @@ export function useConnectWithSiwe() {
     try {
       const { nonce } = await client.auth.initSIWE({ address });
       const SIWEMessage = createSIWEMessage(address, nonce, chainId);
+
       const signature = await signMessage(config, { message: SIWEMessage });
 
       // if has user, we link the wallet
@@ -63,11 +70,21 @@ export function useConnectWithSiwe() {
     }
     catch (err) {
       log("Failed to connect with SIWE", err);
-      if (err instanceof AxiosError) {
-        onError && onError("Failed to connect with SIWE", err.request.status);
+      if (!onError) return;
+
+      let message = err instanceof Error ? err.message : err instanceof AxiosError ? err.message : String(err);
+
+      if (message.includes("User rejected the request.")) {
+        message = "User rejected the request.";
+      } else if (message.includes("Invalid signature")) {
+        message = "Invalid signature. Please try again.";
+      } else if (message.includes("An error occurred when attempting to switch chain")) {
+        message = "Failed to switch chain. Please switch your wallet to the correct network and try again.";
       } else {
-        onError && onError("Failed to connect with SIWE");
+        message = "Failed to connect with SIWE.";
       }
+
+      onError(message, err instanceof AxiosError ? err.request.status : undefined);
     }
   }, [client, user, updateUser, log, address, chainId, config, connector]);